Resurrect junit support (#281)
* Resurrected the RoboVM junit support. This commit just pulls in the relevant
sources from the old repositories (robovm-maven-resolver, robovm-junit and
robovm-maven-plugin) without any changes except for the pom.xml files.

* Resurrected the RoboVM junit support. This commit upgrades the RoboVM surefire
provider to the latest version of surefire (2.21.0) and makes sure everything
compiles and runs as expected. The resolver is upgraded to a newer version of
shrinkwrap (2.2.6) and the resolver tests are fixed up to work as expected.