* fixed different behaviour of resolving list of Java9+ modules in jrt file system between Oracle JDK and OpenJDK. (#411)
On OpenJDK (v13 was tested) when getting path using `Paths.get(URI.create("jrt:/modules"))` its returns it as /modules/modules as result fails to resolve.
It resolves list of modules when using "jrt:/" uri, but it doesn't work if "jrt:/modules" was used before "jrt:/". E.g. this part is heavily broken