* applied 68360-Tidy up locking in OpenSSLSocketImpl that improves multithreading (https://android-review.googlesource.com/c/platform/external/conscrypt/+/68360)
We guard all state with a single lock "stateLock", which
replaces usages of "this" and "handshakeLock". We do not
perform any blocking operations while holding this lock.
In particular, startHandshake is no longer synchronized.
We use a single integer to keep track of handshake state
instead of a pair of booleans.
Also fix a bug in getSession, the previous implementation
wouldn't work in cut-through mode.