* fixed: Undefined symbols for architecture arm64 xxxx[checkcast]
Issue happens when aggressive mode of tree shaker is used.
This caused to opted-out weak methods to be replaced with stubs. If such method refers to class that also was dropped and uses checkcast/ldr wrappers linking will fail as these symbols are in object files that are not included.

The fix: instead of referencing for external symbol weak stub that just throws exception is created.
During linking these weak stubs will be replaced with valid implementation from target files